Adds functionality to admin intructors
Fixes several bugs. links in the views together. Adds ability to edit
instructors

Manikin scaffold, view, integration

Created a scaffold for manikins accessable by admins.

Linked in displaies for manikins in scenario and adding manikins to
scenarios. Manikin.type was a reserved word so changed this to sim_type
with a migration.