Adds views to mailer
Since ActionMailer requires two views for multipart emails this adds a
rake task that will read a partial view writen in markdown and build the
needed text and html views. The html view will render the partial into a
variable then translate it through Maruku. Were as the text view only
renders the partial. This allows erb code to stay in the partial view
and prevents Maruku choking on them. Translating to markdown is done on
the fly. This includes the maruki gem in environment.rb and the
generated files in .gitignore