Fixes bad date print out on tech schedule
The day and date were missing from the tech schedule. This was because
the partial view that was used only took into account each single event
and not the whole week of events. This modifies how you use the partial
view to iterate of the events in a single week.

Fixes bug #194